+++ R/import.R
@@ -69,7 +69,7 @@ import_dataset <- function(project,
                            metadata_colors = NULL,
                            calc_gg_cor = TRUE) {
     verbose <- !is.null(getOption("MCView.verbose")) && getOption("MCView.verbose")
-    verify_project_dir(project)
+    verify_project_dir(project, create = TRUE)
 
     cli_alert_info("Importing {.field {dataset}}")