+++ R/utils_scdb.R
@@ -3,6 +3,8 @@ init_temp_scdb <- function(scdb, matrix, mc, mc2d, network = NULL, dataset, scdb
     fs::dir_create(scdb_dir)
     metacell::scdb_init(base_dir = scdb_dir, force_reinit = TRUE)
 
+    scdb <- normalizePath(scdb)
+
     if (!force_copy && dir.exists(system.file("scdb", package = "MCView"))) {
         scdb_files <- list.files(system.file("scdb", package = "MCView"), full.names = TRUE)
     } else {
@@ -12,5 +14,9 @@ init_temp_scdb <- function(scdb, matrix, mc, mc2d, network = NULL, dataset, scdb
         scdb_files <- fs::path(scdb, files_to_copy)
     }
 
-    purrr::walk(scdb_files, ~ fs::link_create(.x, fs::path(scdb_dir, basename(.x))))
+    purrr::walk(scdb_files, ~ {
+        if (!fs::file_exists(.x)) {
+            fs::link_create(.x, fs::path(scdb_dir, basename(.x)))
+        }
+    })
 }